webscraper-pro/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0
lxml>=4.9.0
fake-useragent>=1.4.0
pytest>=7.4.0
